requirements.txt
django
opencage
allauth
crispy_forms
bootstrap4
multiselectfield
django_filter
pillow